Updated source without useless conditions.
To support commits at ctubio/tribeca,
you can buy-me-a-drink with a small git tip at:
  1GitTipgxvKB3zjCLXRcSgDhC9pivkpc7u

I promise to drink chocolate milk in the development of the next commit.

To request new features or in case this commit breaks something for you,
please create a new github issue with all possible details,
but never share your API Keys!

+++ src/lib/stdev.cc
@@ -12,7 +12,7 @@ namespace tribeca {
     v8::Local<v8::Float64Array> seqD = info[3].As<v8::Float64Array>();
 
     double factor = info[4]->NumberValue();
-    float minTick = info[5]->NumberValue();
+    double minTick = info[5]->NumberValue();
 
     v8::Local<v8::Object> obj = Nan::New<v8::Object>();
